feat: project-manager DeleteProject接口适配多租户

+++ bcs-services/bcs-project-manager/internal/actions/project/delete.go
@@ -15,6 +15,8 @@ package project
 import (
 	"context"
 
+	"github.com/Tencent/bk-bcs/bcs-services/bcs-project-manager/internal/auth"
+	"github.com/Tencent/bk-bcs/bcs-services/bcs-project-manager/internal/config"
 	"github.com/Tencent/bk-bcs/bcs-services/bcs-project-manager/internal/store"
 	"github.com/Tencent/bk-bcs/bcs-services/bcs-project-manager/internal/util/errorx"
 	proto "github.com/Tencent/bk-bcs/bcs-services/bcs-project-manager/proto/bcsproject"
@@ -38,6 +40,17 @@ func NewDeleteAction(model store.ProjectModel) *DeleteAction {
 func (da *DeleteAction) Do(ctx context.Context, req *proto.DeleteProjectRequest) error {
 	da.ctx = ctx
 	da.req = req
+
+	// 不能删除别的租户的 Project
+	project, err := da.model.GetProject(ctx, req.GetProjectID())
+	if err != nil {
+		return errorx.NewDBErr(err.Error())
+	}
+	// 未找到，或者项目属于别的租户，都不能删除
+	// TODO 通过一个数据库操作完成，不用先查
+	if project == nil || (config.GlobalConf.MultiTenantEnabled && project.TenantID != auth.GetTenantFromCtx(ctx)) {
+		return errorx.NewReadableErr(errorx.ProjectNotExistsErr, "project not found")
+	}
 
 	if err := da.model.DeleteProject(ctx, req.ProjectID); err != nil {
 		return errorx.NewDBErr(err.Error())
